% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rap.R
\name{auto}
\alias{auto}
\title{Automatically select the output type}
\usage{
auto()
}
\description{
When \code{auto()} is used as the type on the left hand side of a \code{\link[=rap]{rap()}} formula,
the common type will be automatically deduced using the infrastructure
provided by \code{vctrs}.
}
\examples{
library(dplyr)

# The output for each row is an integer of length 1
# so the type is automatically deduced as integer
starwars \%>\%
  rap(count = auto() ~ length(films))

# auto() is mostly useful for automatically
# determining an atomic common type (integer, double, etc).
# When you have more complicated objects, you have to list() them
starwars \%>\%
  rap(model = auto() ~ list(lm(height ~ mass, starwars)))

# In this case, it's more useful to use the default rap() behavior
# which already assumes the result should be a list
starwars \%>\%
  rap(model = ~lm(height ~ mass, starwars))

}
