% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{blfdr_stat}
\alias{blfdr_stat}
\title{test statistics based on 3-component beta mixture}
\usage{
blfdr_stat(U, eta1, eta2, eta3, eta4, extraParam)
}
\arguments{
\item{U}{A vector of u-values to evaluate the beta density on}

\item{eta1}{a vector of linear predictors, each being \eqn{ (1, X)'\beta_l} for each test. Same length as U.}

\item{eta2}{a vector of linear predictors, each being \eqn{ (1, X)'\beta_r} for each test. Same length as U.}

\item{eta3}{a vector of linear predictors, each being \eqn{ (1, X)'\theta_l} for each test. Same length as U.}

\item{eta4}{a vector of linear predictors, each being \eqn{ (1, X)'\theta_r} for each test. Same length as U.}

\item{extraParam}{A two-element vector, \eqn{(\gamma_l, \gamma_r)}.}
}
\value{
the test statistics which are the accessor functions evaluated at U.
}
\description{
This is a function that evaluate the accessor functions based on the 3-component beta mixture on a vector of u-values
}
\details{
TBD
}
\examples{
TBD
}
