% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createParaMat}
\alias{createParaMat}
\title{create para matrix}
\usage{
createParaMat(param, X_tilde)
}
\arguments{
\item{param}{the beta and theta you use to create a paraMat}

\item{X_tilde}{the augmented covariate matrix}
}
\description{
create para matrix
}
