% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_asymp.R
\name{EM_asymp}
\alias{EM_asymp}
\title{EM for asymptotic ZAP}
\usage{
EM_asymp(
  z,
  param,
  tol = 1e-04,
  pprob = NULL,
  maxit,
  extraParam,
  X_tilde,
  uvals.cutoff
)
}
\description{
an internal function used to perform the EM for zap_asymp
}
\examples{
TBD
}
