% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_asymp.R
\name{EM_asymp_fix_point_fn}
\alias{EM_asymp_fix_point_fn}
\title{EM asymp fix point function for SQUAREM}
\usage{
EM_asymp_fix_point_fn(param, U, extraParam, X_tilde)
}
\description{
an internal function used to perform the EM for zap_asymp using the  SQUAREM package
}
\examples{
TBD
}
