% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_asymp.R
\name{EM_asymp_obj_fn}
\alias{EM_asymp_obj_fn}
\title{Negative log-likelihood of the 3-component beta mixture}
\usage{
EM_asymp_obj_fn(param, U, extraParam, X_tilde)
}
\description{
an internal function used to return the objective function value for the EM of zap_asymp
}
\examples{
TBD
}
