% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_finite.R
\name{EM_finite}
\alias{EM_finite}
\title{EM for finite-sample ZAP}
\usage{
EM_finite(
  U,
  U_mirror,
  tol,
  unmask_set,
  mask_set,
  pprob = NULL,
  maxit,
  beta0,
  piMat,
  extraParam,
  X_tilde
)
}
\description{
an internal function used to perform the EM for zap_finite
}
\examples{
TBD
}
