% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{form_Hmat_unmask_init}
\alias{form_Hmat_unmask_init}
\title{form H matrix from unmasked data with initial pi matrix and beta0}
\usage{
form_Hmat_unmask_init(U, beta0, piMat, X_tilde, extraParam)
}
\arguments{
\item{x}{A single integer.}
}
\description{
form H matrix from unmasked data with initial pi matrix and beta0
}
