% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{form_Hmat_Yab_mask}
\alias{form_Hmat_Yab_mask}
\title{This function forms the Hmat and Ymat for masked u-values ALONE}
\usage{
form_Hmat_Yab_mask(U_mask, U_mirror_mask, paraMat, extraParam)
}
\arguments{
\item{x}{A single integer.}
}
\description{
This function forms the Hmat and Ymat for masked u-values ALONE
}
