% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{form_Hmat_Yab_mask_init}
\alias{form_Hmat_Yab_mask_init}
\title{This function forms the Hmat and Ymat for masked u-values ALONE}
\usage{
form_Hmat_Yab_mask_init(
  U_mask,
  U_mirror_mask,
  beta0,
  piMat,
  X_tilde,
  extraParam
)
}
\arguments{
\item{x}{A single integer.}
}
\description{
This function forms the Hmat and Ymat for masked u-values ALONE
}
