% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{neg_LogLike_LeftBeta}
\alias{neg_LogLike_LeftBeta}
\title{negative log-likelihood of of a left-leaning beta density}
\usage{
neg_LogLike_LeftBeta(beta_l, U, X_tilde, gamma_l)
}
\arguments{
\item{beta_l}{The first shape parameter}

\item{U}{A vector of u-values to evaluate the beta density on}

\item{X_tilde}{A matrix of the covariates augmented with 1's as intercepts. Should have the same number of rows as the length of U}

\item{gamma_l}{The second shape parameter, should be greater than 2}
}
\value{
the evaluated negative log likelihood value
}
\description{
This is a function that evaluate the left leaning beta negative log likelihood for a vector of u-values
}
\details{
TBD
}
\examples{
TBD
}
