% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{neg_LogLike_RightBeta_grad}
\alias{neg_LogLike_RightBeta_grad}
\title{negative right beta log-likelihood gradient}
\usage{
neg_LogLike_RightBeta_grad(beta_r, U, X_tilde, gamma_r)
}
\arguments{
\item{beta_r}{The first shape parameter}

\item{U}{A vector of u-values to evaluate the beta density on}

\item{X_tilde}{A matrix of the covariates augmented with 1's as intercepts. Should have the same number of rows as the length of U}

\item{gamma_r}{The second shape parameter, should be greater than 2}
}
\value{
the evaluated negative log likelihood gradient, a vector of dimension p
}
\description{
This is a function that evaluate the right leaning beta negative log likelihood gradient for a vector of u-values
}
\details{
TBD
}
\examples{
TBD
}
