% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{overall_Hmat}
\alias{overall_Hmat}
\title{arrange an overall Hmat}
\usage{
overall_Hmat(Hmat_mask, Hmat_unmask, mask_set, unmask_set)
}
\arguments{
\item{x}{A single integer.}
}
\description{
arrange an overall Hmat
}
