% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{overall_Ymat}
\alias{overall_Ymat}
\title{arrange an overall Ymat}
\usage{
overall_Ymat(Ymat_mask, U_unmask, mask_set, unmask_set)
}
\arguments{
\item{x}{A single integer.}
}
\description{
arrange an overall Ymat
}
