% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zap-package.R
\docType{package}
\name{zap-package}
\alias{zap}
\alias{zap-package}
\title{zap: z-value based covariate-adaptive testing}
\description{
It does z-value based covariate-assisted multiple two-sided testing with FDR control.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dmhleung/zap}
  \item \url{https://academic.oup.com/jrsssb/article/84/5/1886/7072884}
}

}
\author{
\strong{Maintainer}: Dennis Leung \email{dennis.leung@unimelb.edu.au} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
