% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zap_asymp.R
\name{zap_asymp}
\alias{zap_asymp}
\title{Asymptotic ZAP}
\usage{
zap_asymp(
  z,
  X,
  beta_incpt_init = c(log(5), log(5)),
  theta_incpt_init = c(log(0.1), log(0.1)),
  gamma = c(4, 4),
  alpha = 0.05,
  gen_size = 50000,
  maxit = 100,
  tol = 1e-04,
  ep = 1,
  uvals.cutoff = 1e-15
)
}
\arguments{
\item{z}{a vector of z-values}

\item{X}{a matrix of dimension n by p, where  n is the length of z and p is
the number of covariates. p should at least be 1.}

\item{beta_incpt_init}{a numeric vector of length 2 concatenating initial values
for the INTERCEPT coefficients of the vectors \eqn{\beta_l} and
\eqn{\beta_r} in that order. Default to c(log(5),  log(5)).}

\item{theta_incpt_init}{a numeric vector of length 2 concatenating initial values
for the INTERCEPT coefficients of the vectors \eqn{\theta_l} and
\eqn{\theta_r} in that order. Default to c(log(0.1),  log(0.1)).}

\item{gamma}{a numeric vector of length 2 concatenating the second shape parameters
\eqn{\gamma_l}, \eqn{\gamma_r}  for the left-leaning and right-leaning beta
densities in that order. Default to c(4, 4).}

\item{alpha}{targeted FDR level. Default to 0.05.}

\item{gen_size}{number of random uniformly distributed realizations generated
to estimate the mirror statistics. Default to 50000.}

\item{ep}{additive factor in the numerator of the FDP estimates in ZAP.
Default is 1. Changing it may result in an inflated FDR of the procedure.}

\item{uvals.cutoff}{an ``epsilon-small" number. The u-values will be cut off
to be within the interval \deqn{[uvals.cutoff, 1 - uvals.cutoff].} This
stabilizes the beta mixture density estimation. Default is 1e-15.}
}
\value{
a list with containing the followings: \itemize{ \item rej_index: a
vector of indices corresponding to the rejected hypotheses. Can be of
length 0 if no hypotheses are rejected. \item stat: a vector of resulting
test statistics \item mirror_stat: a vector of "mirror" test statistics
\item param: a vector of estimated parameters concatenating  \eqn{\beta_l}, \eqn{\beta_r}, \eqn{\theta_l} and
\eqn{\theta_r} in that order }
}
\description{
Asymptotic version of ZAP based on a 3-component beta mixture
working model.
}
\details{
TBD
}
\examples{
TBD


}
