% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zap_finite.R
\name{zap_finite}
\alias{zap_finite}
\title{Finite-sample ZAP}
\usage{
zap_finite(
  z,
  X,
  beta_incpt_init = c(log(5), log(5)),
  gamma = c(4, 4),
  s_l = NULL,
  s_r = NULL,
  alpha = 0.05,
  maxit = 10,
  tol = 1e-04,
  nfits = 100,
  uvals.cutoff = 1e-15,
  ms = F,
  x = NULL,
  df = 6
)
}
\arguments{
\item{z}{a vector of z-values}

\item{X}{a matrix of dimension n by p, where  n is the length of z and p is
the number of covariates. p should at least be 1.}

\item{beta_incpt_init}{a numeric vector of length 2 concatenating initial
values for the INTERCEPT coefficients of the vectors \eqn{\beta_l} and
\eqn{\beta_r} in that order. Default to c(log(5),  log(5)).}

\item{gamma}{a numeric vector of length 2 concatenating the second shape parameters
\eqn{\gamma_l}, \eqn{\gamma_r}  for the left-leaning and right-leaning beta
densities in that order. Default to c(4, 4).}

\item{s_l, s_r}{left and right initial thresholding functions, each should be
a vector of the same length as z. If equal to NULL, s_l will default to a
vector with all components equal to 0.2, and s_r will default to a vector
with all components equal to 0.8.}

\item{alpha}{targeted FDR level.  Default to 0.05.}

\item{maxit}{maximum number of iterations for every EM update}

\item{tol}{tolerance}

\item{nfits}{maximum number of EM updates during the ZAP procedure. Default to 100.}

\item{uvals.cutoff}{an ``epsilon-small" number. The u-values will be cut off
to be within the interval \deqn{[uvals.cutoff, 1 - uvals.cutoff].} This
stabilizes the beta mixture density estimation. Default is 1e-15.}

\item{ms}{indicator of whether model selection should be performed. Default
to FALSE. Under construction, please don't touch this option for now.}

\item{x}{Under construction, please don't touch this option for now.}

\item{df}{Under construction, please don't touch this option for now.}
}
\value{
a vector of indices of the rejected hypotheses
}
\description{
Finite-sample version of ZAP based on a 3-component beta mixture
working model.
}
\details{
TBD
}
\examples{
#TBD
}
