% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zap.R
\name{zap_opts}
\alias{zap_opts}
\title{Create options list for writing data}
\usage{
zap_opts(
  transform,
  verbosity,
  lgl,
  int,
  fct,
  dbl,
  str,
  list,
  lgl_threshold,
  int_threshold,
  fct_threshold,
  dbl_threshold,
  str_threshold,
  dbl_fallback,
  ...
)
}
\arguments{
\item{transform}{Enable transformations? Default: TRUE.  Setting to 
FALSE will disable all transformations.}

\item{verbosity}{Verbosity level. Default: 0 (no text output).
\describe{
  \item{64}{Return a data.frame with information on each SEXP within the object. 
       \code{start} and \code{end} values are the position of the object within
       the \emph{uncompressed} stream}
}}

\item{lgl}{transformation method for logical vectors. Default: 'packed'
\describe{
  \item{\code{raw}}{Raw. No transformation}
  \item{\code{packed}}{Packed 2 bits per logical value}
}}

\item{int}{transformation method for integer vectors. Default: 'deltaframe'
\describe{
  \item{\code{raw}}{Raw. No transformation}
  \item{\code{zzshuf}}{Zig-zag encoding, delta and shuffle}
  \item{\code{deltaframe}}{Delta frame-of-reference coding}
}}

\item{fct}{transformation method for factors vectors. Default: 'packed'
\describe{
  \item{\code{raw}}{Raw. No transformation}
  \item{\code{packed}}{Packed minimal bits per level}
}}

\item{dbl}{transformation method for doubles (and complex) vectors. Default: 'alp'
\describe{
  \item{\code{raw}}{Raw. No transformation}
  \item{\code{shuffle}}{Byte shuffle}
  \item{\code{delta_shuffle}}{Byte shuffle with delta}
  \item{\code{alp}}{ALP, Adaptive Lossless Floating Point compression}
}}

\item{str}{transformation method for character vectors. Default: 'mega'
\describe{
  \item{\code{raw}}{Raw. No transformation}
  \item{\code{mega}}{Concatenate all strings.  Length implicitly encoded by null bytes in strings}
  \item{\code{dict}}{Use unique strings within the vector as a dictionary, and 
        encode strings as integer indexes into the dictionary}
}}

\item{list}{transformation method for lists (and data.frames).  Default: 'raw'
\describe{
  \item{\code{raw}}{Raw. All lists written out in-full}
  \item{\code{reference}}{Cache lists and data.frames as they are seen, and 
        if seen again, write out a reference to the prior object rather than 
        writing out in-full.}
}}

\item{int_threshold, lgl_threshold, fct_threshold, dbl_threshold, str_threshold}{Below this threshold, no transformation will be done. All default to
0, meaning transformation is always attempted.}

\item{dbl_fallback}{if \code{dbl = 'alp'}, the data is not
always conducive to this compression scheme and after probing the
data the code can exit early and try a different method.  
The \code{dbl_fallback} variable nominates the fallback method if ALP
transformation is being attempted, but fails. The options are the
same as for the \code{dbl} argument (excluding option \code{'alp'})}

\item{...}{expert level options}
}
\value{
named list
}
\description{
Create options list for writing data
}
\examples{
myopts <- zap_opts(dbl = 'shuffle')
zap_write(seq(1:1000) * 1.5, opts = myopts)
}
