% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zap.R
\name{zap_read}
\alias{zap_read}
\title{Unserialize R object from raw vector or file}
\usage{
zap_read(src, opts = list(), ...)
}
\arguments{
\item{src}{Serialization source - either a raw vector of filename.}

\item{opts}{Named list of options.   See \code{\link{zap_opts}()}}

\item{...}{other named options to be included in \code{opts}. See
\code{\link{zap_opts}()} for list of valid options.}
}
\value{
Unserialized R object
}
\description{
Unserialize R object from raw vector or file
}
\examples{
raw_vec <- zap_write(head(mtcars))
head(raw_vec, 50)
length(raw_vec)
zap_read(raw_vec)
}
