% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zap.R
\name{zap_version}
\alias{zap_version}
\title{Get version of internal transformation code.}
\usage{
zap_version()
}
\value{
Integer
}
\description{
This version number is the same as the version number present in the
header of the serialized data.
}
\examples{
zap_version()
}
