% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zap.R
\name{zap_write}
\alias{zap_write}
\title{Serialize R object to raw vector or file}
\usage{
zap_write(
  x,
  dst = NULL,
  compress = Sys.getenv("zap_compress_default"),
  opts = list(),
  ...
)
}
\arguments{
\item{x}{R object}

\item{dst}{Serialization destination. Default: NULL means to return the raw vector.
If a character string is given it is assumed to be the path
to the output file.}

\item{compress}{compression type. Default: 'zstd' if available, otherwise 'gzip'.
This is set in the 'zap_compress_default' environment variable after
being detected during package start.
Other valid values 'none', 
'xz', 'bzip2'.  Compression is done using \code{memCompress()}}

\item{opts}{Named list of options.   See \code{\link{zap_opts}()}}

\item{...}{other named options to be included in \code{opts}. See
\code{\link{zap_opts}()} for list of valid options.}
}
\value{
IF \code{dst} is NULL, then return a raw vector, otherwise data
        is written to file and nothing is returned.
}
\description{
Serialize R object to raw vector or file
}
\examples{
raw_vec <- zap_write(head(mtcars))
head(raw_vec, 50)
length(raw_vec)
zap_read(raw_vec)
}
