% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{form_Hmat_unmask}
\alias{form_Hmat_unmask}
\title{form H matrix from unmasked data}
\usage{
form_Hmat_unmask(U, paraMat, extraParam)
}
\arguments{
\item{x}{A single integer.}
}
\description{
form H matrix from unmasked data
}
